frame-close mode
this mode places one frame inside each call to encode() and then flushes
the frame out entirely. the purpose of this mode is to create buffers
which can be played with gaps between them -- e.g. if the encoder faces
garbage collection or other pauses

todo: create a planner which can place a limit on the maximum frame
size. the current implementation will silently allow frames to cross
buffer boundaries.