blocking mode, closable encoder/decoder
this is a large patch that makes it possible to perform blocking reads
and writes in quiet. it also creates a complete and sensible close() and
new return values which can allow quiet to exist in a pipeline, such as
with portaudio. this patch does change return values slightly and builds
a more robust error signalling mechanism